@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace OnlineShop.Models
+﻿namespace OnlineShop.Models
 {
     public class Item
     {
