@@ -15,16 +15,16 @@
 </div>
 <div id="homeBody">    
     <aside>
-        <p>Categories</p>
+        <p> &nbsp; Categories</p>
         <ul>
             <li>
-                <a asp-area="" asp-controller="Home" asp-action="Category" asp-route-category="Laptops and computers"><span>+</span> Laptops and computers</a>
+                <a asp-area="" asp-controller="Home" asp-action="Category" asp-route-category="Laptops and computers"> &nbsp; Laptops and computers</a>
             </li>
             <li>
-                <a asp-area="" asp-controller="Home" asp-action="Category" asp-route-category="Smartphones"><span>+</span> Smartphones</a>
+                <a asp-area="" asp-controller="Home" asp-action="Category" asp-route-category="Smartphones"> &nbsp; Smartphones</a>
             </li>
-            <li><span>+</span> TV</li>
-            <li><span>+</span> Electronics</li>
+            <li> &nbsp; TV</li>
+            <li> &nbsp; Electronics</li>
         </ul>
     </aside>
     <div class="main">
